Move query function to query interface and fix more duplication in the chainstate interface
